zipkin translation: move attributes to internal package
