+++ translator/trace/zipkinv2/to_translator.go
@@ -33,7 +33,7 @@ import (
 	"go.opentelemetry.io/collector/internal/occonventions"
 	"go.opentelemetry.io/collector/translator/conventions"
 	tracetranslator "go.opentelemetry.io/collector/translator/trace"
-	"go.opentelemetry.io/collector/translator/trace/zipkin"
+	"go.opentelemetry.io/collector/translator/trace/internal/zipkin"
 )
 
 var _ model.ToTracesTranslator = (*ToTranslator)(nil)
